в функциях IPC_virtAlloc() и heapAlloc() тип аргумента заменен с int на size_t

+++ include/sysipc.h
@@ -68,9 +68,9 @@ GIPCY_API    int IPC_getPrivateProfileString( const IPC_str *lpAppName, const IP
 
 GIPCY_API    int IPC_writePrivateProfileString( const IPC_str *lpAppName, const IPC_str *lpKeyName, const IPC_str *lpString, const IPC_str *lpFileName );
 
-GIPCY_API    void* IPC_heapAlloc(int nSize);
+GIPCY_API    void* IPC_heapAlloc(size_t nSize);
 GIPCY_API    int IPC_heapFree(void *ptr);
-GIPCY_API    void* IPC_virtAlloc(int nSize);
+GIPCY_API    void* IPC_virtAlloc(size_t nSize);
 GIPCY_API    int IPC_virtFree(void *ptr);
 
 GIPCY_API    long IPC_interlockedDecrement(volatile long *val );
